refactor: extract installer helpers into focused modules
Move module-scope helper functions out of the 1,015-line installer.ts
into four focused modules: toml-helpers.ts, settings-merge.ts,
json-helpers.ts, and install-helpers.ts. The SymbiontInstaller class
and its public API (syncSkillSymlinks, MYCO_MCP_SERVER_NAME) stay in
installer.ts with imports from the new modules.
